refactor: #21 병합 충돌 해결
